[ADD] db user delete functions
