+++ libs/database/database.h
@@ -14,6 +14,7 @@
 #include "fields/thread.h"
 #include "fields/discussion.h"
 #include "llist.h"
+#include "search/database_search_functions.h"
 
 typedef struct database_s
 {
@@ -49,7 +50,10 @@ database_t *db_destuction(database_t *database);
  *
  * @return A boolean value.
  */
-bool db_add_user(database_t *db, user_t *user);
+static inline bool db_add_user(database_t *db, user_t *user)
+{
+    return llist_append(db->users, user);
+};
 
 /**
  * @brief It searches the database for a user with the given id
@@ -59,7 +63,10 @@ bool db_add_user(database_t *db, user_t *user);
  *
  * @return A pointer to a user_t struct.
  */
-user_t *db_search_user_by_id(database_t *db, const id_t id);
+user_t *db_search_user_by_id(database_t *db, const id_t id)
+{
+    return llist_search(db->users, &user_id_compare, id);
+};
 
 /**
  * @brief It searches the database for a user with the given pseudo
@@ -69,4 +76,33 @@ user_t *db_search_user_by_id(database_t *db, const id_t id);
  *
  * @return A pointer to the user_t struct.
  */
-user_t *db_search_user_by_pseudo(database_t *db, const char *pseudo);
+user_t *db_search_user_by_pseudo(database_t *db, const char *pseudo)
+{
+    return llist_search(db->users, &user_pseudo_compare, pseudo);
+};
+
+/**
+ * @brief It deletes a user, with the given id, in the database
+ *
+ * @param db The database to search in.
+ * @param id The id of the user to delete.
+ *
+ * @return A boolean value.
+ */
+bool db_delete_user_by_id(database_t *db, const id_t id)
+{
+    return llist_delif(db->users, &user_id_compare, id);
+}
+
+/**
+ * @brief It deletes a user, with the given pseudo, in the database
+ *
+ * @param db the database.
+ * @param pseudo the pseudo of the user to delete.
+ *
+ * @return A boolean value.
+ */
+bool db_search_user_by_pseudo(database_t *db, const char *pseudo)
+{
+    return llist_delif(db->users, &user_pseudo_compare, pseudo);
+}
