+++ libs/database/search/user_search.c
@@ -13,6 +13,7 @@ bool user_id_compare(const void *elem, const void *search)
 {
     const user_t *u = (const user_t *)elem;
     const id_t *id = (const id_t *)id;
+
     if (u->id == *id)
         return true;
     return false;
@@ -22,6 +23,7 @@ bool user_pseudo_compare(const void *elem, const void *search)
 {
     const user_t *u = (const user_t *)elem;
     const char *pseudo = (const char *)pseudo;
+
     if (strcmp(u->pseudo, pseudo) == 0)
         return true;
     return false;
